Extract p0 and rho0 from amsca to the property of BulkSolver and BulkAlgorithm.  These two properties are global scalars for a solver object. Create a BulkProperty class to hold the rest properties bulk (K), eta (I don't know what it is used for), dynamic viscosity (mu), density (rho), and velocity (vel; a 3-vector).
This CS resolves an issue that the solver runs the density to negative.  The
root cause is the BulkNonrefl.init() method of the BC class didn't calculate
the ghost geometry for the conservation element (center).
sc_bulk_ghostgeom_mirror_*() subroutines are added.

I also found that this BulkNonrefl BC implementation shouldn't use the taylor
expansion from the old solution, or it gives wrong answer and explode.  This
behavior should be documented after the CESE method is documented.