Ref issue #150 Get rid of _MethodList form solvcon.solver
`_MethodList` is a slightly modified `list` that helps tracking the names of time-marching methods in a derived class of `solvcon.solver.MeshSolver` (aka `solvcon.MeshSolver`).  But it's not a straight-forward design.  In the subclass, a `_MethodList` needs to be instantiated as a class attribute (by calling `MeshSolver.new_method_list()`) as `_MMNAMES`.  A time-marching method needs to be decorated by `_MMNAMES.register`.

By following [the trick used by Django](http://blog.hirokiky.org/2013/11/04/getting_the_definition_order_of_class_attributes_in_python.html), we can simplify the design.  Instead of using a special list object, we let `_MMNAMES` be an ordinary `tuple`.  Each time-marching method is decorated by `MeshSolver.register_marcher` (instead of the decorator `_MMNAMES.register`, which doesn't exist anymore), which adds an attribute to track the order where the method is defined in the class.  The metaclass `MeshSolverMeta` extracts the order mark and create the `_MMNAMES` tuple.