Port the unstructured mesh data structure from Python/C to C++.
This is a major part of the work porting the whole system to be C++-inside-Python-outside.  Unstructured mesh definition is the core of the SOLVCON architecture.  The core was developed using Python to manage the resources, and C and Cython wrappers were made to speed up slow parts.  After this porting, we switch to use C++ to manage the resources, and use pybind11 to provide the wrapper for Python to manipulate the data.

All unit tests and functional tests (ftests/*) remain working.  Very little is changed in the tests.  Pickling for the C++ objects are added to the tests.

The ported facility is march::mesh::UnstructuredBlock class template.  It supports both 2 and 3 dimensional spaces.  In the Python land a wrapping class solvcon.block.UnstructuredBlock is made to dynamically dispatch to either of the 2 or 3 dimensional class.  The old solvcon.block.Block class is moved to solvcon.block_legacy.LegacyBlock.  solvcon.block.Block becomes an alias to solvcon.block.UnstructuredBlock.

The following C functions were ported for the C++ UnstructuredBlock class with the minimum modification (i.e., they are purely C style still):

1. sc_mesh_extract_faces_from_cells() -> UnstructuredBlock::build_faces_from_cells()
2. sc_mesh_build_ghost() -> UnstructuredBlock::fill_ghost()
3. sc_mesh_calc_metric() -> UnstructuredBlock::calc_metric()
4. sc_mesh_build_rcells() -> UnstructuredBlock::build_rcells()
5. sc_mesh_build_csr() -> UnstructuredBlock::build_csr()

Also include cdoe to call external graph partitioner.  Currently hardcode to use SCOTCH / SCOTCHMETIS.

For now, sign-compare warning/error is turned off in both CMakeLists.txt and setup.py.  It should be turned on when the legacy C is ported to modern C++.

To support UnstructuredBlock, LookupTableCore is made copyable.  The copy constructor and assignment operator do "cheap" copy only, that is, copy only the metadata of the LookupTableCore, not the buffer content.  BoundaryData is also made copyable.  Pickle support is added to BoundaryData and UnstructuredBlock.  Note, pybind11 support only version 2 pickle protocol.

Other changes:

- In Travis-CI, provide verbose output for libmarch building and nosetests output.
- Fix Buffer destructure null pointer check.
- Include CONDA_PREFIX when building with conda.