Manage UnstructuredBlock using shared pointer.
Make UnstructuredBlock uncopyable and unmoveable as it is now managed by a shared pointer.

pybind11 seems to need to initialize an object using its reference, or I haven't found a way not to.  My current workaround is to allow a loophole in UnstructuredBlock constructor so that the wrapper can do placement new.

Remove sc_mesh_t as it is not used at all.