Add march::Vector and march::Matrix types to simplify Cartesian vector analysis.
Although there’s still room for optimization, the Vector and Matrix types are efficient for the calculations needed in the multi-dimensional CESE method.

The Vector class template supports =, +=, -=, *=, /=, +, -, *, / operators, and dot, square, length methods.  cross product is implemented as a free function so that it would work in both two- and three-dimensional space.

The Matrix class template supports =, +=, -=, *=, /=, +, -, *, / operators, and free functions product, volume, unnormalized_inverse.