Port code for ideal gas dynamics solver to C++.
The code compiles and runs, but the solution hasn't been verified at all and I have a good faith that the result is incorrect.  Use sandbox/gasplus/obrf/go as the test script to show the code runs.  The ported code is exposed to Python via march.gas sub module.  The porting include:

- Core CESE solver.
- Boundary condition treatments.
- Physical quantity calculation.

march::gas::Quantity is part of Solver class.  Do not use shared pointers to manage Quantity class.  Match the physics variable list (`aname`) to the fields in Quality class.  Remove the useless PhysicsAnchor _varlist_.

In the new C++ system, the boundary condition treatments is called "Trim".  The ported treatments include:

- No-op for debugging.
- Non-reflective
- Slip wall
- Inlet.