Move the holder of Python-derived Anchors to C++ (ref issue #197)
Add `march::gas::AnchorChain::LifeManager` (internal class).  The life manager is responsible for owning resources needed by the `Anchor` instances.  It is polymorphic and in the Python (pybind11) wrapper it is extended by `PythonAnchorManager` to own the Python-derived part of the `Anchor` instances.  Thus, the hack done in `solvcon.hook.MeshHook._deliver_anchor` becomes obsolete and is removed.

A cyclic-referencing bug is found in `solvcon.parcel.gasplus.inout.CflAnchor`.  To fix it, the CFL statistics is moved to C++ (it was stored in a Python dictionary).

`march::InstanceCounter` class template is created and enabled for `march::gas::Solver` to help debugging resource leak.